requirements.txt
asgiref==3.5.2
Django==2.1.15
django-cors-headers==3.13.0
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
djangorestframework==3.14.0
pyodbc==4.0.34
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
